version 3

+++ model/oceanformer.py
@@ -329,6 +329,6 @@ class Xuanming(nn.Module):
 if __name__=="__main__":
     model = Xuanming()
     x = torch.randn(2, 19, 30, 400, 441)
-    x_mark = torch.Tensor([[11, 30], [1,28]]).long()
-    y_mark = torch.Tensor([[11, 30], [1,28]]).long()
+    x_mark = torch.Tensor([[11, 30], [1,28]]).float()
+    y_mark = torch.Tensor([[11, 30], [1,28]]).float()
     y = model(x, x_mark, y_mark)
